+++ src/prelude.rs
@@ -16,6 +16,8 @@ It exports the following symbols:
 - Judge and Return traits
 - `tear!`, `terror!` and `twist!` macros
 - The useful `tear_if!` and `anybox!` macros
+- `next_if!` and `last_if` because they're unlikely to conflict
+- (f=experimental) `impl_judge_from_try!`
 */
 
 pub use crate::ValRet::{self, *};
@@ -28,3 +30,6 @@ pub use crate::Return;
 // Macros
 pub use crate::{tear, terror, twist};
 pub use crate::{tear_if, anybox};
+pub use crate::{next_if, last_if};
+
+#[cfg(feature = "experimental")] pub use crate::impl_judge_from_try;